requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
arrow==1.1.0
binaryornot==0.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
fastapi==0.63.0
h11==0.9.0
haruka-bot==1.0.7
httpcore==0.12.3
httpx==0.16.1
idna==2.10
importlib-metadata==2.1.1
Jinja2==3.0.1
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
nb-cli==0.3.2
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0a10
packaging==20.9
Pillow==8.2.0
poyo==0.5.0
prompt-toolkit==1.0.14
pydantic==1.7.3
pyee==8.1.0
pyfiglet==0.8.post1
Pygments==2.9.0
pygtrie==2.4.2
PyInquirer==1.0.3
pyparsing==2.4.7
pyppeteer==0.2.5
python-dateutil==2.8.1
python-dotenv==0.17.1
python-slugify==5.0.2
pytz==2021.1
qrcode==6.1
regex==2021.4.4
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.13.6
text-unidecode==1.3
tinydb==4.4.0
tqdm==4.60.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
uvicorn==0.11.8
wcwidth==0.2.5
websockets==8.1
win32-setctime==1.0.3
zipp==3.4.1
